[REF] Separate contact tokens from 'tokenCompatSubscriber'
This just moves code around - loads of testing - but it separates
the contact entity stuff from the 'compat' - presumably 'compatibility'
subscriber.
